+++ FBSDKCoreKit/FBSDKCoreKit/AppEvents/Internal/FBSDKAppEventsUtility.m
@@ -319,7 +319,7 @@ static NSArray<NSString *> *standardEvents;
     // If there's an logging override app id present, then we don't want to use the client token since the client token
     // is intended to match up with the primary app id (and AppEvents doesn't require a client token).
     NSString *clientTokenString = [FBSDKSettings clientToken];
-    if (clientTokenString && appID && [appID isEqualToString:token.appID]) {
+    if (clientTokenString && appID && ([appID isEqualToString:token.appID] || token == nil)) {
       tokenString = [NSString stringWithFormat:@"%@|%@", appID, clientTokenString];
     } else if (appID) {
       tokenString = nil;